Start sketching out web-connectivity test
