+++ src/nettests/web-connectivity/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "ooni-web-connectivity",
+  "version": "0.3.0",
+  "description": "Official OONI test for running Web Connectivity",
+  "dependencies": {
+    "measurement-kit": "0.1.0"
+  }
+}
